+++ test/tracks.js
@@ -1,0 +1,80 @@
+const assert = require('assert');
+const createLocalTracks = require('../lib/createlocaltrack');
+
+const isFirefox = navigator.userAgent.indexOf("Firefox") > 0;
+
+['audio', 'video'].forEach(kind => {
+  const createLocalTrack = createLocalTracks[kind];
+  const description = 'Local' + kind[0].toUpperCase() + kind.slice(1) + 'Track';
+
+  describe(description, () => {
+    let localTrack = null;
+
+    beforeEach(() => {
+      return createLocalTrack().then(_localTrack => {
+        localTrack = _localTrack;
+      });
+    });
+
+    afterEach(() => {
+      localTrack.stop();
+      localTrack = null;
+    });
+
+    describe('.isStopped', () => {
+      context('before calling #stop', () => {
+        it('.isStopped is false', () => {
+          assert(!localTrack.isStopped);
+        });
+      });
+
+      context('after calling #stop', () => {
+        beforeEach(() => {
+          localTrack.stop();
+        });
+
+        it('.isStopped is true', () => {
+          assert(localTrack.isStopped);
+        });
+      });
+
+      context('when the underlying MediaStreamTrack ends', () => {
+        beforeEach(() => {
+          localTrack.mediaStreamTrack.stop();
+        });
+
+        it('.isStopped is true', () => {
+          assert(localTrack.isStopped);
+        });
+      });
+    });
+
+    describe('"stopped" event', () => {
+      let stoppedEvent = null;
+
+      beforeEach(() => {
+        stoppedEvent = new Promise(resolve => {
+          localTrack.once('stopped', resolve);
+        });
+      });
+
+      afterEach(() => {
+        stoppedEvent = null;
+      });
+
+      context('when #stop is called', () => {
+        it('emits "stopped"', () => {
+          localTrack.stop();
+          return stoppedEvent;
+        });
+      });
+
+      context('when the underlying MediaStreamTrack ends', () => {
+        (isFirefox ? it.skip : it)('emits "stopped"', () => {
+          localTrack.mediaStreamTrack.stop();
+          return stoppedEvent;
+        });
+      });
+    });
+  });
+});